+++ Clipster/Clipster/Domain/Protocol/UseCase/URL/SanitizeURLUseCase.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol SanitizeURLUseCase {
+    func execute(urlString: String) -> Result<URL, URLValidationError>
+}
