+++ Clipster/Clipster/Domain/UseCase/URL/DefaultParseURLUseCase.swift
@@ -7,19 +7,15 @@ final class DefaultParseURLUseCase: ParseURLUseCase {
         self.urlRepository = urlMetaRepository
     }
 
-    func execute(urlString: String) async -> Result<(URLMetadata?, Bool), URLValidationError> {
-        guard let sanitizeURL = try? sanitizeURL(urlString: urlString).get() else {
-            return .failure(.badURL)
-        }
-
-        let resolveFinalURL = await urlRepository.resolveRedirectURL(initialURL: sanitizeURL)
+    func execute(url: URL) async -> Result<(URLMetadata?, Bool), URLValidationError> {
+        let resolveFinalURL = await urlRepository.resolveRedirectURL(initialURL: url)
 
         let htmlResult = await urlRepository.fetchHTML(from: resolveFinalURL)
 
         switch htmlResult {
         case .success(let html):
             let screenshotData = await urlRepository.captureScreenshot(rect: nil)
-            let parsedMetadata = createParsedURLMetadata(url: sanitizeURL, html: html, screenshotData: screenshotData)
+            let parsedMetadata = createParsedURLMetadata(url: url, html: html, screenshotData: screenshotData)
             return .success((parsedMetadata, true))
 
         case .failure(let error):
@@ -29,36 +25,6 @@ final class DefaultParseURLUseCase: ParseURLUseCase {
 }
 
 extension DefaultParseURLUseCase {
-    func sanitizeURL(urlString: String) -> Result<URL, URLValidationError> {
-        let lowercased = urlString.lowercased()
-        let correctedURLString: String
-
-        if lowercased.hasPrefix("https://") || lowercased.hasPrefix("http://") {
-            correctedURLString = urlString
-        } else if lowercased.hasPrefix("https:/") || lowercased.hasPrefix("http:/") {
-            correctedURLString = lowercased.replacingOccurrences(of: "https:/", with: "https://")
-                .replacingOccurrences(of: "http:/", with: "https://")
-        } else if lowercased.hasPrefix("https:")  || lowercased.hasPrefix("http:") {
-            correctedURLString =  lowercased.replacingOccurrences(of: "https:", with: "https://")
-                .replacingOccurrences(of: "http:", with: "http://")
-        } else {
-            correctedURLString = "https://\(urlString)"
-        }
-
-        guard correctedURLString != "https://" || correctedURLString != "http://" else {
-            return .failure(.badURL)
-        }
-
-        guard let url = URL(string: correctedURLString) else {
-            return .failure(.badURL)
-        }
-
-        guard let host = url.host, host.contains(".") else {
-            return .failure(.badURL)
-        }
-        return .success(url)
-    }
-
     func createParsedURLMetadata(url: URL, html: String, screenshotData: Data?) -> URLMetadata {
         let ogTitle = extractOGContent(html: html, property: "og:title")
         let title = ogTitle ?? extractHTMLTagContent(html: html, property: "title") ?? "제목 없음"
